merge

+++ ZC_IT_TimeTracking/GetGoalRuleDetails_Result.cs
@@ -14,8 +14,8 @@ namespace ZC_IT_TimeTracking
     public partial class GetGoalRuleDetails_Result
     {
         public int Goal_RuleID { get; set; }
-        public double Performance_RangeFrom { get; set; }
-        public double Performance_RangeTo { get; set; }
+        public int Performance_RangeFrom { get; set; }
+        public int Performance_RangeTo { get; set; }
         public double Rating { get; set; }
     }
 }
