+++ ZC_IT_TimeTracking/Models/GoalRule.cs
@@ -7,8 +7,8 @@ namespace ZC_IT_TimeTracking.Models
 {
     public class GoalRule
     {
-        public float RangeFrom { get; set; }
-        public float RangeTo { get; set; }
-        public float Rating { get; set; }
+        public int RangeFrom { get; set; }
+        public int RangeTo { get; set; }
+        public int Rating { get; set; }
     }
 }
